+++ lib/src/layer/tile_layer/level.dart
@@ -1,0 +1,11 @@
+import 'package:flutter_map/flutter_map.dart';
+
+class Level {
+  final CustomPoint origin;
+  final double zoom;
+
+  Level({
+    required this.origin,
+    required this.zoom,
+  });
+}
