+++ lib/src/layer/tile_layer/tile_provider/file_tile_provider_web.dart
@@ -1,15 +1,13 @@
-import 'dart:io';
-
 import 'package:flutter/widgets.dart';
-
-import 'package:flutter_map/src/layer/tile_layer.dart';
-import 'package:flutter_map/src/layer/tile_provider/tile_provider.dart';
+import 'package:flutter_map/flutter_map.dart';
 
 /// FileTileProvider
+
 class FileTileProvider extends TileProvider {
   const FileTileProvider();
+
   @override
   ImageProvider getImage(Coords<num> coords, TileLayerOptions options) {
-    return FileImage(File(getTileUrl(coords, options)));
+    return NetworkImage(getTileUrl(coords, options));
   }
 }
